73 taxonomy pruning

+++ dap_aria_mapping/pipeline/make_topic_name_assignments.py
@@ -20,20 +20,19 @@ Returns:
 """
 import argparse, json, boto3, time, subprocess
 from toolz import pipe
-from dap_aria_mapping import logger, PROJECT_DIR, BUCKET_NAME
 from functools import partial
 from nesta_ds_utils.loading_saving.S3 import get_bucket_filenames_s3
+from dap_aria_mapping import logger, PROJECT_DIR, BUCKET_NAME, chatgpt_args
 from dap_aria_mapping.getters.taxonomies import (
     get_cooccurrence_taxonomy,
     get_semantic_taxonomy,
     get_topic_names,
 )
+from nesta_ds_utils.loading_saving.S3 import upload_obj
 from dap_aria_mapping.getters.openalex import get_openalex_works, get_openalex_entities
 from dap_aria_mapping.utils.entity_selection import get_sample, filter_entities
 from dap_aria_mapping.utils.topic_names import *
 from dap_aria_mapping.utils.chatgpt import revChatGPTWrapper, webChatGPTWrapper
-
-OUTPUT_DIR = PROJECT_DIR / "outputs" / "interim" / "topic_names"
 
 
 def save_names(
@@ -83,7 +82,7 @@ if __name__ == "__main__":
     parser.add_argument(
         "--name_type",
         nargs="+",
-        help="The type of names to use. Can be 'entity', 'journal', or both.",
+        help="The type of names to use. Can be 'entity', 'journal', revChatGPT, or webChatGPT.",
         default="entity",
     )
 
@@ -104,6 +103,12 @@ if __name__ == "__main__":
         help="The number of articles to use. If -1, all articles are used.",
     )
 
+    parser.add_argument(
+        "--postproc",
+        help="Whether to use the postprocessed chatgpt names.",
+        action="store_true",
+    )
+
     parser.add_argument("--save", help="Whether to save the plot.", action="store_true")
 
     parser.add_argument(
@@ -118,7 +123,7 @@ if __name__ == "__main__":
     logger.info("Loading data - taxonomy")
     taxonomies = []
     if "cooccur" in args.taxonomy:
-        cooccur_taxonomy = get_cooccurrence_taxonomy()
+        cooccur_taxonomy = get_cooccurrence_taxonomy(postproc=args.postproc)
         taxonomies.append(["cooccur", cooccur_taxonomy])
     if "centroids" in args.taxonomy:
         semantic_centroids_taxonomy = get_semantic_taxonomy("centroids")
@@ -159,10 +164,14 @@ if __name__ == "__main__":
                     clust_counts, num_entities=args.n_top, show_count=args.show_count
                 )
                 if args.save:
+                    if args.postproc:
+                        name_type = "entity_postproc"
+                    else:
+                        name_type = "entity"
                     logger.info("Saving dictionary as pickle")
                     save_names(
                         taxonomy=taxlabel,
-                        name_type="entity",
+                        name_type=name_type,
                         level=level,
                         n_top=args.n_top,
                         names=names,
@@ -187,17 +196,32 @@ if __name__ == "__main__":
                     )
 
             if "chatgpt" in args.name_type[0].lower():
-                entity_names = get_topic_names(
-                    taxonomy_class=taxlabel,
-                    name_type="entity",
-                    level=level,
-                    n_top=args.n_top,
-                )
+                logger.info("Loading ChatGPT names")
+                if args.postproc:
+                    entity_names = get_topic_names(
+                        taxonomy_class=taxlabel,
+                        name_type="entity_postproc",
+                        level=level,
+                        n_top=args.n_top,
+                    )
+                else:
+                    entity_names = get_topic_names(
+                        taxonomy_class=taxlabel,
+                        name_type="entity",
+                        level=level,
+                        n_top=args.n_top,
+                    )
 
-                files_with_name = get_bucket_filenames_s3(
-                    bucket_name=BUCKET_NAME,
-                    dir_name=f"outputs/topic_names/class_{taxlabel}_nametype_chatgpt_top_{args.n_top}_level_{level}.json",
-                )
+                if not args.postproc:
+                    files_with_name = get_bucket_filenames_s3(
+                        bucket_name=BUCKET_NAME,
+                        dir_name=f"outputs/topic_names/class_{taxlabel}_nametype_chatgpt_top_{args.n_top}_level_{level}.json",
+                    )
+                else:
+                    files_with_name = get_bucket_filenames_s3(
+                        bucket_name=BUCKET_NAME,
+                        dir_name=f"outputs/topic_names/postproc/level_{level}.json",
+                    )
 
                 if len(files_with_name) > 0:
                     logger.info("ChatGPT names already exist")
@@ -206,6 +230,7 @@ if __name__ == "__main__":
                         name_type="chatgpt",
                         level=level,
                         n_top=args.n_top,
+                        postproc=args.postproc,
                     )
 
                     num_topics_file, num_topics_total = (
@@ -263,15 +288,50 @@ if __name__ == "__main__":
                     logger.info(f"Selecting random chatbot to use")
 
                     if args.name_type[0] == "revChatGPT":
-                        chatbot_num = np.random.randint(1, 8)
+                        chatbot_num = np.random.randint(
+                            1, len(chatgpt_args["TOKENS"]) + 1
+                        )
                         logger.info(
                             f"Starting batch of topics using chatbot {chatbot_num}"
                         )
-
-                    random_sample = random.sample(list(entity_names.keys()), 6)
+                    try:
+                        random_sample = list(entity_names.keys())[:6]
+                    except:
+                        random_sample = entity_names.keys()
                     sample_entities = [
                         (topic, entity_names[topic]) for topic in random_sample
                     ]
+
+                    chunk_str = "\n\n ".join(
+                        ["List " + ': "'.join(x) + '"' for x in sample_entities]
+                    )
+
+                    first_query = (
+                        chatgpt_args["QUESTION"]
+                        + " \n\n "
+                        + f"{chunk_str}"
+                        + " \n\n "
+                        + chatgpt_args["CLARIFICATION"]
+                        + " \n\n"
+                        + chatgpt_args["EXAMPLE"]
+                        + " \n\n"
+                        + chatgpt_args["REQUEST"]
+                    )
+
+                    routine_query = (
+                        chatgpt_args["NEXT"]
+                        + f"\n\n {chunk_str} \n\n"
+                        + chatgpt_args["CLARIFICATION"]
+                        + "\n\n"
+                        + chatgpt_args["EXAMPLE"]
+                    )
+
+                    error_query = (
+                        chatgpt_args["ERROR"]
+                        + "\n\n"
+                        + chatgpt_args["EXAMPLE"]
+                        + "\n\n Please try again. Only return the list in the required structure. \n\n"
+                    )
 
                     tries = 0
                     while True:
@@ -282,13 +342,17 @@ if __name__ == "__main__":
                                 chatgpt_names = revchatgpt(
                                     chatbot_num=chatbot_num,
                                     chatgpt_names=chatgpt_names,
-                                    sample_entities=sample_entities,
+                                    first_query=first_query,
+                                    routine_query=routine_query,
+                                    error_query=error_query,
                                 )
                             elif args.name_type[0] == "webChatGPT":
                                 chatgpt_names = webchatgpt(
                                     chatgpt_names=chatgpt_names,
-                                    sample_entities=sample_entities,
                                     tries=tries,
+                                    first_query=first_query,
+                                    routine_query=routine_query,
+                                    error_query=error_query,
                                 )
 
                             # Update number of topics with names & pending groups
@@ -299,13 +363,20 @@ if __name__ == "__main__":
                             }
                             if args.save:
                                 logger.info("Saving dictionary as json")
-                                save_names(
-                                    taxonomy=taxlabel,
-                                    name_type="chatgpt",
-                                    level=level,
-                                    n_top=args.n_top,
-                                    names=chatgpt_names,
-                                )
+                                if not args.postproc:
+                                    save_names(
+                                        taxonomy=taxlabel,
+                                        name_type="chatgpt",
+                                        level=level,
+                                        n_top=args.n_top,
+                                        names=chatgpt_names,
+                                    )
+                                else:
+                                    upload_obj(
+                                        chatgpt_names,
+                                        bucket=BUCKET_NAME,
+                                        path_to=f"outputs/topic_names/postproc/level_{str(level)}.json",
+                                    )
                             break
 
                         except Exception as e:
@@ -313,15 +384,14 @@ if __name__ == "__main__":
                             logger.info(
                                 f"ChatGPT failed to respond. Reason: {e}. Trying again."
                             )
-                            time.sleep(np.random.randint(6, 12))
+                            time.sleep(np.random.randint(20, 30))
                             if tries > 3:
                                 logger.info(
                                     "ChatGPT failed to respond. Idling for 5-10 minutes."
                                 )
-                                import subprocess
+                                if args.name_type[0] == "webChatGPT":
+                                    subprocess.run("pkill firefox", shell=True)
 
-                                # pkill firefox using subprocess
-                                subprocess.run("pkill firefox", shell=True)
                                 time.sleep(np.random.randint(300, 600))
 
                                 logger.info("Restarting chatbot")
